+++ convert_onnx.py
@@ -1,22 +1,11 @@
 import os
 import torch
 import argparse
-from fairseq.models.transformer_super import TransformerSuperModel
 from fairseq import tasks, options, utils
 from fairseq.data import Dictionary, indexed_dataset
 from fairseq.models import fairseq_model
-
-class WrapperModel(torch.nn.Module):
-    def __init__(self, model):
-        super(WrapperModel, self).__init__()
-        self.model = model
-
-    def prepare_for_onnx_export(self):
-        if hasattr(self.model, "prepare_for_onnx_export_"):
-            self.model.prepare_for_onnx_export_()
-
-    def forward(self, src_tokens, src_lengths, prev_output_tokens):
-        return self.model(src_tokens=src_tokens, src_lengths=src_lengths, prev_output_tokens=prev_output_tokens)
+from fairseq.models.transformer_super import TransformerSuperModel
+from wrapperModels import WrapperModelONNX
 
 def generate_dummy_data(args):
     # specify the length of the dummy input for profile
@@ -40,7 +29,7 @@ def generate_dummy_data(args):
 
 
 def export_to_onnx(model, src_tokens, src_lengths, prev_output_tokens, dataset_name):
-    wrapper_model = WrapperModel(model)
+    wrapper_model = WrapperModelONNX(model)
     wrapper_model.prepare_for_onnx_export()
 
     onnx_file_path = f"./onnx_models/{dataset_name}.onnx"
@@ -71,12 +60,16 @@ def main():
     print(f"| Buildng model {args.arch}...")
     task = tasks.setup_task(args)
     model = task.build_model(args)
-    model.eval()
 
     src_tokens, src_lengths, prev_output_tokens = generate_dummy_data(args)
+    # src_tokens = src_tokens.detach()
+    # src_lengths = src_lengths.detach()
+    # prev_output_tokens = prev_output_tokens.detach()
 
     config_sam = utils.sample_configs(utils.get_all_choices(args), reset_rand_seed=False, super_decoder_num_layer=args.decoder_layers)
     model.set_sample_config(config_sam)
+    model.eval()
+
 
     print("| Exporting model to ONNX...")
     export_to_onnx(
