+++ errors/convert_onnx_wmt14.en-de.txt
@@ -88,26 +88,3 @@ Columns 1 to 6 5.2654e-02 -2.5555e-02  6.3955e-02  2.6899e-02  2.1934e-02 -2.166
  1.9818e-02 -3.5226e-02 -3.1740e-02 -7.2057e-02 -4.6456e-02  4.1625e-03
 -6.0681e-02 -4.5354e-02 -2.9682e-02 -2.3449e-02 -3.0763e-02 -1.3436e-02
 생략
-
-Columns 7 to 12-6.7885e-02 -6.1797e-02  1.7018e-02  4.9302e-02 -5.2939e-02  6.3248e-02
--1.2946e-02 -6.5291e-02 -2.6385e-02 -5.3994e-02 -5.2861e-02 -3.9288e-02
--3.3564e-02  3.2547e-02  4.9065e-02 -6.3903e-02 -4.2359e-02  6.9462e-02
- 6.2701e-02  5.6702e-02  3.8288e-02  6.7552e-02 -5.4527e-02 -3.5672e-02
- 6.4061e-02 -2.4346e-02 -3.9133e-02 -3.2286e-04  1.5823e-02  6.2137e-02
-생략
-
-Columns 13 to 18-6.0902e-02 -4.5265e-02 -4.9637e-02 -3.4135e-02 -6.0008e-02  4.5968e-02
--4.1238e-02 -6.2319e-02  1.4529e-02  4.5366e-04  4.0897e-02 -4.9999e-03
--6.0120e-02 -2.0127e-02 -6.3787e-02 -2.1085e-02 -4.0123e-02  1.5304e-02
- 5.3828e-02  4.5800e-02 -6.8371e-02  1.6286e-02  5.8270e-02 -2.8063e-02
-생략
-
-Columns 19 to 24-2.9598e-02 -8.8776e-03 -7.6874e-04 -7.2088e-02 -4.0058e-02  4.3920e-02
- 4.2529e-02 -6.3252e-02 -6.4053e-02 -5.8259e-02  1.9061e-02 -9.2618e-03
--1.8214e-02 -1.6179e-02 -1.6896e-02  1.7970e-02  5.3538e-02  3.0584e-02
- 2.3226e-02 -1.4189e-02 -7.3636e-03  6.4690e-02 -3.2685e-02 -4.2817e-02
- 5.6030e-02  5.6524e-02  1.1277e-02  3.2686e-02  4.8328e-02 -2.0387e-02
--5.3174e-02 -3.4038e-02 -6.5118e-02  5.2382e-02 -8.9708e-03 -3.7703e-02
--4.8935e-02  4.5438e-02  2.6109e-02 -4.6482e-03 -3.9024e-02 -2.3799e-02
-~
-
